refactor(marazion): split lake/stream from_bounds into build helpers
Thin the leaf facades so path/budget/shelf layout stays separate from depression+apron assembly. Share apron salt groups and depth jitter so lake and stream call the same noise plumbing.

- `stream/{path,build}.rs` — endpoints/grades vs corridor stamp parts
- `lake/{budget,shelf,build}.rs` — inscribed budget, centroid/shelf survey, bowl+apron parts
- `ApronNoiseSalts::{LAKE,STREAM}` + `jittered_depth` in apron module

## Agent Dialogue

- **Developer:** Increase water bleed across lake/stream (prior).
- **Agent:** Bumped fill bleed/undercut defaults.
- **Developer:** Break from_bounds into submodule helpers; unify/simplify where sensible.
- **Agent:** Split lake and stream into layout vs build modules; shared apron salts and depth jitter.

- **Developer:** Prefer modern lake.rs + lake/ layout over mod.rs.
- **Agent:** Moved lake/mod.rs → lake.rs and stream/mod.rs → stream.rs; submodules stay under lake/ and stream/.
