+++ maybraid/durham/marazion/src/lake.rs
@@ -9,36 +9,24 @@
 //! clearance), with a small noisy rotation. Leaves must still leave room for
 //! rim + apron outside the water body.
 
+mod budget;
+mod build;
+mod shelf;
+
+pub use budget::LakeBandBudget;
+pub use shelf::shelf_base_height;
+
 use crate::apron::WatershedApronParams;
-use crate::complex::{WatershedApronShelf, WatershedDepressionComplex};
-use crate::depression::{WatershedDepression, WatershedDepressionKind};
-use crate::fill::{WaterFill, WaterSurface};
-use crate::noise::{n01_at, n01_freq, n11_at, scale_noise_freq};
-use bevy_math::Vec2;
-use jersey_terrain_stamps::{
-	EllipseRegion, JerseyModulation, Region2D, RegionBowlModulation, RegionNoise,
+use crate::complex::WatershedDepressionComplex;
+use crate::fill::WaterFill;
+use crate::lake::build::{build_parts, LakeLayout};
+use crate::lake::shelf::{
+	aspect_u01, planned_center as planned_center_impl, rim_width_u01, rotation_u11, shelf_levels,
+	water_scale_u01,
 };
+use bevy_math::Vec2;
+use jersey_terrain_stamps::JerseyModulation;
 use procedural_common::Bounds2;
-
-/// Minimum water half-axis (world units); smaller budgets skip the stamp.
-const MIN_WATER_RADIUS: f32 = 8.0;
-
-/// Salt for per-leaf water-radius undershoot.
-const WATER_SCALE_SALT: u32 = 0x1A7E_512E;
-/// Salt for per-leaf rim-width undershoot.
-const RIM_WIDTH_SALT: u32 = 0x1A7E_71B7;
-/// Salt for aspect blend (circle ↔ full leaf aspect).
-const ASPECT_SALT: u32 = 0x1A7E_A5EC;
-/// Salt for ellipse rotation.
-const ROTATION_SALT: u32 = 0x1A7E_207A;
-/// Salt for per-leaf rim height amplitude draw.
-const RIM_HEIGHT_AMP_SALT: u32 = 0x1A7E_A17A;
-/// Salt for per-leaf rim height frequency draw.
-const RIM_HEIGHT_FREQ_SALT: u32 = 0x1A7E_F7E9;
-/// Salt for per-leaf apron indent amplitude draw.
-const APRON_AMP_SALT: u32 = 0x1A7E_A70A;
-/// Salt for per-leaf apron frequency draw.
-const APRON_FREQ_SALT: u32 = 0x1A7E_AF7E;
 
 /// Authoring knobs for a Marazion lake stamp.
 ///
@@ -153,7 +141,7 @@ impl Default for LakeParams {
 
 			water_sink: 0.9,
 			rim_lift: 1.25,
-			terrain_undercut: 1.75,
+			terrain_undercut: 2.5,
 			shelf_amp: 2.0,
 			shelf_sample_count: 6,
 
@@ -164,246 +152,14 @@ impl Default for LakeParams {
 			island_lift: 5.5,
 
 			shore_indent_frac: 0.18,
-			// Authored at ref; √ scale + min(1). Slightly below the old linear-era
-			// knobs so high-pass effective Hz stay near the look that worked, while
-			// sub-ref ponds no longer inherit an amplified harshness.
 			shore_freq: 0.022,
 
 			apron: WatershedApronParams::default(),
 
-			rim_bleed_frac: 0.35,
-			shore_fade: 2.0,
+			rim_bleed_frac: 0.5,
+			shore_fade: 3.0,
 		}
 	}
-}
-
-/// Elliptical band budget derived from per-axis clearance at the lake centroid.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct LakeBandBudget {
-	/// Water half-axes in the ellipse local frame.
-	pub water_radii: Vec2,
-	pub rotation: f32,
-	pub rim_width: f32,
-	pub apron_width: f32,
-	/// Plateau half-axes (`water + rim`) in the local frame.
-	pub plateau_radii: Vec2,
-	pub mu: f32,
-}
-
-impl LakeBandBudget {
-	/// Characteristic (short) water half-axis.
-	pub fn water_radius(&self) -> f32 {
-		self.water_radii.min_element()
-	}
-
-	/// Characteristic (short) plateau half-axis.
-	pub fn plateau_radius(&self) -> f32 {
-		self.plateau_radii.min_element()
-	}
-
-	/// Budget water / rim / apron from inscribed clearance at `center`.
-	///
-	/// Rim + apron claim first (isotropic widths); water takes a noisy fraction
-	/// of the leftover on each axis, blended between circular and full leaf
-	/// aspect. A small rotation is applied and axes are rescaled so the rotated
-	/// ellipse AABB still fits. `water_u01` / `rim_u01` / `aspect_u01` /
-	/// `rotation_u11` should be stable per leaf.
-	///
-	/// Returns `None` when the leaf cannot host a meaningful three-band lake.
-	pub fn try_inscribed(
-		bounds: Bounds2,
-		center: Vec2,
-		params: LakeParams,
-		water_u01: f32,
-		rim_u01: f32,
-		aspect_u01: f32,
-		rotation_u11: f32,
-	) -> Option<Self> {
-		let min = bounds.min;
-		let max = bounds.max;
-		let room = Vec2::new(
-			(center.x - min.x).min(max.x - center.x).max(0.0),
-			(center.y - min.y).min(max.y - center.y).max(0.0),
-		);
-		let short_room = room.min_element();
-		if short_room < MIN_WATER_RADIUS * 2.0 {
-			return None;
-		}
-		let mu = params.mu.min(short_room * 0.2).max(0.0);
-		let available = (room - Vec2::splat(mu)).max(Vec2::ZERO);
-		let short_avail = available.min_element();
-		if short_avail < MIN_WATER_RADIUS * 2.0 {
-			return None;
-		}
-
-		// Enforce leaf clearance ≥ 2 · body diameter on the short axis.
-		let max_water_short = (short_room * 0.5).min(short_avail * 0.45);
-		let rim_claim = (short_avail * params.rim_frac.clamp(0.05, 0.45))
-			.max(short_avail * 0.12)
-			.min(short_avail * 0.42);
-		let rim_hi = 1.0;
-		let rim_lo = params.rim_width_min.clamp(0.2, rim_hi);
-		let rim = rim_claim * (rim_lo + (rim_hi - rim_lo) * rim_u01.clamp(0.0, 1.0));
-		let apron = (short_avail * params.apron_frac.max(0.05))
-			.max(short_avail * 0.14)
-			.min(short_avail * 0.72);
-
-		// Per-axis leftover after isotropic rim+apron claim. Cap the short axis
-		// so leaf ≥ 2× body; the long axis may use more of its clearance.
-		let leftover = Vec2::new(
-			(available.x - rim_claim - apron).max(0.0),
-			(available.y - rim_claim - apron).max(0.0),
-		);
-		let long_frac = params.long_axis_frac.clamp(0.45, 0.95);
-		let leftover = Vec2::new(
-			if available.x <= available.y {
-				leftover.x.min(max_water_short)
-			} else {
-				leftover.x.min(available.x * long_frac)
-			},
-			if available.y <= available.x {
-				leftover.y.min(max_water_short)
-			} else {
-				leftover.y.min(available.y * long_frac)
-			},
-		);
-
-		let size_hi = params.water_scale.clamp(0.05, 1.0);
-		let size_lo = params.water_scale_min.clamp(0.05, size_hi);
-		let size_frac = size_lo + (size_hi - size_lo) * water_u01.clamp(0.0, 1.0);
-
-		let circ = leftover.min_element() * size_frac;
-		let full = leftover * size_frac;
-		let aspect = aspect_blend(params, short_room, aspect_u01);
-		let mut water = Vec2::new(circ + (full.x - circ) * aspect, circ + (full.y - circ) * aspect);
-		if water.min_element() < MIN_WATER_RADIUS {
-			return None;
-		}
-		// Re-check 2× body vs short leaf room.
-		if 2.0 * short_room < 2.0 * (2.0 * water.min_element()) * 0.99 {
-			return None;
-		}
-
-		let rotation = rotation_u11.clamp(-1.0, 1.0) * params.rotation_amp.max(0.0);
-		// Shrink so rotated water+rim+apron AABB fits inscribed clearance.
-		let band = rim + apron;
-		let fit_radii = water + Vec2::splat(band);
-		let (s, c) = rotation.sin_cos();
-		let aabb = Vec2::new(
-			((fit_radii.x * c).abs().powi(2) + (fit_radii.y * s).abs().powi(2)).sqrt(),
-			((fit_radii.x * s).abs().powi(2) + (fit_radii.y * c).abs().powi(2)).sqrt(),
-		);
-		let scale = (available / aabb.max(Vec2::splat(1e-3))).min_element().clamp(0.0, 1.0);
-		water *= scale;
-		if water.min_element() < MIN_WATER_RADIUS {
-			return None;
-		}
-
-		Some(Self {
-			water_radii: water,
-			rotation,
-			rim_width: rim,
-			apron_width: apron,
-			plateau_radii: water + Vec2::splat(rim),
-			mu,
-		})
-	}
-
-	/// Axis-aligned helper used by unit tests (centroid = leaf center, no rotation).
-	pub fn try_from_short_half(
-		short_half: f32,
-		params: LakeParams,
-		water_u01: f32,
-		rim_u01: f32,
-	) -> Option<Self> {
-		let s = short_half.max(0.0);
-		let bounds = Bounds2::from_xz(-s, -s, s, s);
-		Self::try_inscribed(bounds, Vec2::ZERO, params, water_u01, rim_u01, 0.0, 0.0)
-	}
-}
-
-/// Aspect blend ∈ `[0, 1]` — weak on small leaves, strong on large.
-///
-/// ```text
-/// t = 1 - 1/sqrt(max(short_room/ref, 1))   // 0 at ref, →1 as size grows
-/// strength = aspect_strength * lerp(aspect_small, 1, t)
-/// floor    = aspect_floor * t
-/// blend    = strength * lerp(floor, 1, aspect_u01)
-/// ```
-fn aspect_blend(params: LakeParams, short_room: f32, aspect_u01: f32) -> f32 {
-	let u = aspect_u01.clamp(0.0, 1.0);
-	let ratio = (short_room / params.aspect_scale_ref.max(1.0)).max(1.0e-3);
-	// Geometric size factor: flat (0) while small_room ≤ ref, then √ ramp.
-	let t = (1.0 - 1.0 / ratio.max(1.0).sqrt()).clamp(0.0, 1.0);
-	let small = params.aspect_small.clamp(0.0, 1.0);
-	let strength = params.aspect_strength.clamp(0.0, 1.0) * (small + (1.0 - small) * t);
-	let floor = (params.aspect_floor.clamp(0.0, 0.95) * t).clamp(0.0, 0.9);
-	(strength * (floor + (1.0 - floor) * u)).clamp(0.0, 1.0)
-}
-
-/// Per-leaf water-scale unit sample (shared by centroid planning and stamp build).
-fn water_scale_u01(seed: u32, leaf_min: Vec2, params: LakeParams) -> f32 {
-	n01_freq(seed, WATER_SCALE_SALT, leaf_min, params.water_scale_freq)
-}
-
-/// Per-leaf rim-width unit sample (shared by centroid planning and stamp build).
-fn rim_width_u01(seed: u32, leaf_min: Vec2, params: LakeParams) -> f32 {
-	n01_freq(seed, RIM_WIDTH_SALT, leaf_min, params.rim_width_freq)
-}
-
-fn aspect_u01(seed: u32, leaf_min: Vec2) -> f32 {
-	n01_at(seed, ASPECT_SALT, leaf_min)
-}
-
-fn rotation_u11(seed: u32, leaf_min: Vec2) -> f32 {
-	n11_at(seed, ROTATION_SALT, leaf_min)
-}
-
-fn ellipse_region(center: Vec2, radii: Vec2, rotation: f32) -> Region2D {
-	Region2D::Ellipse(EllipseRegion { center, radii: radii.max(Vec2::splat(1e-3)), rotation })
-}
-
-/// Mid-rim characteristic radius used when surveying surrounding terrain.
-fn shelf_survey_radius(budget: &LakeBandBudget) -> f32 {
-	(budget.water_radius() + budget.rim_width * 0.5).max(1.0)
-}
-
-/// Median of `samples` (in-place select). Empty → `0.0`.
-fn median_f32(samples: &mut [f32]) -> f32 {
-	let n = samples.len();
-	if n == 0 {
-		return 0.0;
-	}
-	let mid = n / 2;
-	let (_, med, _) = samples.select_nth_unstable_by(mid, |a, b| a.total_cmp(b));
-	*med
-}
-
-/// Sample pre-watershed height around `center` to set the shelf / rim base.
-///
-/// With `count ≤ 1` (or no sampler), returns the centroid height. Otherwise
-/// takes the **median** of `count` evenly spaced samples on a ring at
-/// `sample_radius` so a single spike or dip at the center does not own `W`.
-pub fn shelf_base_height(
-	center: Vec2,
-	sample_radius: f32,
-	count: u8,
-	height_at: Option<&dyn Fn(f32, f32) -> f32>,
-) -> f32 {
-	let Some(f) = height_at else {
-		return 0.0;
-	};
-	let n = usize::from(count.max(1));
-	if n == 1 || sample_radius <= 1e-3 {
-		return f(center.x, center.y);
-	}
-	let mut hs = Vec::with_capacity(n);
-	for i in 0..n {
-		let ang = (i as f32) * std::f32::consts::TAU / (n as f32);
-		let p = center + Vec2::new(ang.cos(), ang.sin()) * sample_radius;
-		hs.push(f(p.x, p.y));
-	}
-	median_f32(&mut hs)
 }
 
 /// Lake stamp products for one pocket-water leaf.
@@ -431,45 +187,11 @@ pub struct Lake {
 impl Lake {
 	/// Jittered lake centroid used by [`Self::from_bounds`] (for shelf survey).
 	pub fn planned_center(bounds: Bounds2, seed: u32, params: LakeParams) -> Vec2 {
-		let min = bounds.min;
-		let max = bounds.max;
-		let cell_c = Vec2::new((min.x + max.x) * 0.5, (min.y + max.y) * 0.5);
-		let half = Vec2::new((max.x - min.x) * 0.5, (max.y - min.y) * 0.5);
-		let short_half = half.x.min(half.y).max(1.0);
-		// Conservative outer pad from a centered circular budget so the clamp
-		// box stays valid before the inscribed elliptical budget is known.
-		let u = water_scale_u01(seed, min, params);
-		let rim_u = rim_width_u01(seed, min, params);
-		let Some(probe) = LakeBandBudget::try_from_short_half(short_half, params, u, rim_u) else {
-			return cell_c;
-		};
-		let shore_amp = probe.water_radius() * params.shore_indent_frac.clamp(0.0, 0.45);
-		let apron_indent = params
-			.apron
-			.indent_frac_min
-			.max(params.apron.indent_frac_max)
-			.clamp(0.0, 0.5);
-		let apron_amp = probe.apron_width * apron_indent;
-		let outer = probe.plateau_radius() + probe.apron_width + shore_amp.max(apron_amp);
-		let lo = min + Vec2::splat(outer.max(probe.mu));
-		let hi = max - Vec2::splat(outer.max(probe.mu));
-		let ox = n11_at(seed, 0x1A7E_C001, min) * params.centroid_jitter * short_half;
-		let oz = n11_at(seed, 0x1A7E_C002, min) * params.centroid_jitter * short_half;
-		Vec2::new(
-			(cell_c.x + ox).clamp(lo.x.min(hi.x), lo.x.max(hi.x)),
-			(cell_c.y + oz).clamp(lo.y.min(hi.y), lo.y.max(hi.y)),
-		)
+		planned_center_impl(bounds, seed, params)
 	}
 
-	/// Build a three-band lake, or an empty stamp when the leaf is too small.
-	pub fn from_bounds(
-		bounds: Bounds2,
-		seed: u32,
-		params: LakeParams,
-		height_at: Option<&dyn Fn(f32, f32) -> f32>,
-	) -> Self {
-		let min = bounds.min;
-		let empty = |center: Vec2| Self {
+	fn empty(bounds: Bounds2, seed: u32, center: Vec2) -> Self {
+		Self {
 			bounds,
 			seed,
 			center,
@@ -483,8 +205,17 @@ impl Lake {
 			water_level: 0.0,
 			modulations: Vec::new(),
 			fills: Vec::new(),
-		};
+		}
+	}
 
+	/// Build a three-band lake, or an empty stamp when the leaf is too small.
+	pub fn from_bounds(
+		bounds: Bounds2,
+		seed: u32,
+		params: LakeParams,
+		height_at: Option<&dyn Fn(f32, f32) -> f32>,
+	) -> Self {
+		let min = bounds.min;
 		let center = Self::planned_center(bounds, seed, params);
 		let u = water_scale_u01(seed, min, params);
 		let rim_u = rim_width_u01(seed, min, params);
@@ -493,128 +224,43 @@ impl Lake {
 		let Some(budget) =
 			LakeBandBudget::try_inscribed(bounds, center, params, u, rim_u, asp, rot)
 		else {
-			return empty(Vec2::new(
-				(bounds.min.x + bounds.max.x) * 0.5,
-				(bounds.min.y + bounds.max.y) * 0.5,
-			));
+			return Self::empty(
+				bounds,
+				seed,
+				Vec2::new(
+					(bounds.min.x + bounds.max.x) * 0.5,
+					(bounds.min.y + bounds.max.y) * 0.5,
+				),
+			);
 		};
 
-		let anchor = min;
-		let water_r = budget.water_radii;
-		let plateau_r = budget.plateau_radii;
-		let rim_w = budget.rim_width;
-		let apron_w = budget.apron_width.max(1.0);
-		let rotation = budget.rotation;
-		let short_water = budget.water_radius();
-
-		let base_h = shelf_base_height(
+		let levels = shelf_levels(seed, min, center, &budget, params, height_at);
+		let layout = LakeLayout {
 			center,
-			shelf_survey_radius(&budget),
-			params.shelf_sample_count,
-			height_at,
-		);
-		let shelf_anchor = base_h + n11_at(seed, 0x1A7E_50F1, anchor) * params.shelf_amp;
-		let water_level = shelf_anchor - params.water_sink.max(0.0);
-		let rim_level = shelf_anchor + params.rim_lift.max(0.0);
-		let depth = params.depth * (0.65 + 0.7 * n01_at(seed, 0x1A7E_DE07, anchor));
-		let bowl_fade = (rim_w * 0.25).max(0.5).min(short_water * 0.2);
-
-		let rim_bleed = rim_w * params.rim_bleed_frac.max(0.0);
-		let fill_r = water_r + Vec2::splat(rim_bleed);
-		// Keep fill inside plateau+apron on each local axis.
-		let max_fill = plateau_r + Vec2::splat(apron_w - 0.5);
-		let fill_r = Vec2::new(fill_r.x.min(max_fill.x), fill_r.y.min(max_fill.y)).max(water_r);
-		let fill_fade = params.shore_fade.max(1.0);
-
-		let plateau_region = ellipse_region(center, plateau_r, rotation);
-		let water_region = ellipse_region(center, water_r, rotation);
-		let fill_region = ellipse_region(center, fill_r, rotation);
-
-		// ── Size-scaled noise (authored freqs @ NOISE_FREQ_REF_RADIUS) ─────
-		// Shore / apron / rim / bed all use the short water axis so relative
-		// lobe counts stay consistent from low-pass ponds to high-pass lakes.
-		let shore_amp = (short_water * params.shore_indent_frac.clamp(0.0, 0.45))
-			.min(rim_w * 0.85)
-			.max(0.01);
-		let shore_freq =
-			scale_noise_freq(params.shore_freq, short_water, params.apron.noise_freq_power);
-		let shore_noise = RegionNoise::from_seed(seed.wrapping_add(5), shore_freq, shore_amp);
-
-		let apron_noise = params.apron.sample_noise(
+			budget,
+			levels,
+		};
+		let parts = build_parts(seed, min, params, &layout);
+		let compiled = WatershedDepressionComplex::from_lake_node(
+			bounds,
 			seed,
-			anchor,
-			apron_w,
-			short_water,
-			APRON_AMP_SALT,
-			APRON_FREQ_SALT,
-			RIM_HEIGHT_AMP_SALT,
-			RIM_HEIGHT_FREQ_SALT,
-		);
-		let apron_outer = apron_w + apron_noise.apron_amp;
-
-		let depth_noise_freq =
-			scale_noise_freq(params.depth_noise_freq, short_water, params.apron.noise_freq_power);
-		let depth_noise = RegionNoise::from_seed(
-			seed.wrapping_add(9),
-			depth_noise_freq,
-			params.depth_noise_amp.max(0.0),
-		);
-
-		let undercut = params.terrain_undercut.max(0.0);
-		let bed_ceiling = (rim_level + params.island_lift.max(0.0))
-			.max(water_level + undercut + params.depth_noise_amp.max(0.0) * 0.85);
-		let center_bed = water_level - depth;
-		let shore_bed = water_level;
-		let bowl = JerseyModulation::Bowl(
-			RegionBowlModulation::new(
-				water_region.clone(),
-				center_bed,
-				shore_bed,
-				bed_ceiling,
-				params.depth_falloff_power,
-				bowl_fade,
-			)
-			.with_boundary_noise(shore_noise.clone())
-			.with_bed_noise(depth_noise),
-		);
-
-		let fill = WaterFill {
-			region: fill_region,
-			inner_radius: 0.0,
-			outer_radius: fill_fade,
-			noise: Some(shore_noise),
-			surface: WaterSurface::Flat { level: water_level },
-			terrain_undercut: undercut,
-		};
-
-		let depression = WatershedDepression::new(
-			WatershedDepressionKind::LakeBowl,
-			water_region,
-			vec![bowl],
-			Some(fill),
-		);
-		let apron = WatershedApronShelf::LakeFlatten {
-			region: plateau_region,
-			rim_level,
-			outer_radius: apron_outer,
-			apron_noise: apron_noise.apron,
-			rim_height: apron_noise.rim_height,
-		};
-		let compiled =
-			WatershedDepressionComplex::from_lake_node(bounds, seed, depression, apron).compile();
+			parts.depression,
+			parts.apron,
+		)
+		.compile();
 
 		Self {
 			bounds,
 			seed,
-			center,
-			water_radii: water_r,
-			rotation,
-			water_radius: short_water,
-			plateau_radius: budget.plateau_radius(),
-			rim_width: rim_w,
-			apron_width: budget.apron_width,
-			fill_radius: fill_r.min_element(),
-			water_level,
+			center: layout.center,
+			water_radii: layout.budget.water_radii,
+			rotation: layout.budget.rotation,
+			water_radius: layout.budget.water_radius(),
+			plateau_radius: layout.budget.plateau_radius(),
+			rim_width: layout.budget.rim_width,
+			apron_width: layout.budget.apron_width,
+			fill_radius: parts.fill_radius,
+			water_level: layout.levels.water_level,
 			modulations: compiled.modulations,
 			fills: compiled.fills,
 		}
@@ -632,6 +278,9 @@ impl Lake {
 #[cfg(test)]
 mod tests {
 	use super::*;
+	use crate::fill::WaterFill;
+	use crate::lake::budget::{aspect_blend, MIN_WATER_RADIUS};
+	use crate::noise::scale_noise_freq;
 	use bevy_math::Vec2;
 
 	fn apply_mods(mods: &[JerseyModulation], h: f32, x: f32, z: f32) -> f32 {
@@ -666,7 +315,6 @@ mod tests {
 		params.shelf_amp = 0.0;
 		params.shelf_sample_count = 6;
 		let center = Lake::planned_center(bounds, 11, params);
-		// Flat ring at 40; deep spike only at the centroid.
 		let height = |x: f32, z: f32| {
 			let d = Vec2::new(x, z).distance(center);
 			if d < 2.0 {
@@ -713,7 +361,6 @@ mod tests {
 			LakeBandBudget::try_from_short_half(short_half, params, 0.0, 1.0).expect("small");
 		assert!(full.water_radius() > mid.water_radius());
 		assert!(mid.water_radius() > small.water_radius());
-		// Apron stays claimed; size variation is water-only.
 		assert!((full.apron_width - small.apron_width).abs() < 1e-4);
 		assert!(small.water_radius() + 1e-3 >= MIN_WATER_RADIUS);
 		Ok(())
@@ -782,7 +429,6 @@ mod tests {
 			large > small + 0.15,
 			"large-leaf aspect should exceed small-leaf: small={small} large={large}"
 		);
-		// Small leaves stay near aspect_small * strength * u (no floor).
 		let small_hi = aspect_blend(params, params.aspect_scale_ref * 0.5, 1.0);
 		assert!(
 			small_hi < params.aspect_strength * 0.45,
@@ -838,7 +484,6 @@ mod tests {
 		params.rotation_amp = 0.0;
 		params.aspect_strength = 0.0;
 		let lake = Lake::from_bounds(bounds, 11, params, Some(&|_, _| base));
-		// Sit outward of max shore expand so we stay on the shelf, not the bowl.
 		let shore_amp = lake.water_radius * params.shore_indent_frac;
 		let mid_r = lake.water_radius + shore_amp + lake.rim_width * 0.55;
 		let p = lake.center + Vec2::new(mid_r, 0.0);
@@ -904,7 +549,6 @@ mod tests {
 		params.aspect_floor = 0.0;
 		let lake = Lake::from_bounds(bounds, 11, params, Some(&|_, _| base));
 		let mut raised = false;
-		// Near-shore rings: base bed is closer to W so amp can crest above it.
 		for &frac in &[0.55_f32, 0.72, 0.85] {
 			for i in 0..48 {
 				let ang = i as f32 * std::f32::consts::TAU / 48.0;
@@ -984,7 +628,6 @@ mod tests {
 		let bounds = Bounds2::from_xz(0.0, 0.0, 40.0, 200.0);
 		let _ = Lake::planned_center(bounds, 3, LakeParams::default());
 		let lake = Lake::from_bounds_default(bounds, 3);
-		// May be empty (too narrow) but must not panic.
 		let _ = lake.center;
 		Ok(())
 	}
